Node server can now handle using web tokens to verify user emails
Using the provided link, the node server can extract the userId that needs to be verified and set it to verified. After doing so, user can now properly login. Server also returns a web token alongside the user profile when logging in to be used for canActivate plugin in angular - Plugin prevents access to webpages if you aren't verified to be a logged on user.